% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{get_app}
\alias{get_app}
\title{Retrieve a background Shiny application}
\usage{
get_app(port)
}
\arguments{
\item{port}{TCP port the app is listening on (e.g., 3000)}
}
\description{
Retrieve a background Shiny application
}
\seealso{
Other app management functions: 
\code{\link{kill_all_apps}()},
\code{\link{kill_app}()},
\code{\link{list_apps}()}
}
\concept{app management functions}
