% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{kill_all_apps}
\alias{kill_all_apps}
\title{Kill all background Shiny applications}
\usage{
kill_all_apps()
}
\description{
Kill all background Shiny applications
}
\seealso{
Other app management functions: 
\code{\link{get_app}()},
\code{\link{kill_app}()},
\code{\link{list_apps}()}
}
\concept{app management functions}
