% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{kill_app}
\alias{kill_app}
\title{Kill a background Shiny application}
\usage{
kill_app(port)
}
\arguments{
\item{port}{TCP port the app is listening on (e.g., 3000)}
}
\description{
Kill a background Shiny application
}
\seealso{
Other app management functions: 
\code{\link{get_app}()},
\code{\link{kill_all_apps}()},
\code{\link{list_apps}()},
\code{\link{view_app}()}
}
\concept{app management functions}
