% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{list_apps}
\alias{list_apps}
\title{List background Shiny applications}
\usage{
list_apps()
}
\value{
a list of \code{\link[callr:r_process]{callr::r_process}} objects named for the port the Shiny
app is listening on.
}
\description{
List background Shiny applications
}
\seealso{
Other app management functions: 
\code{\link{get_app}()},
\code{\link{kill_all_apps}()},
\code{\link{kill_app}()},
\code{\link{view_app}()}
}
\concept{app management functions}
