% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBackgroundApp.R
\name{runBackgroundApp}
\alias{runBackgroundApp}
\title{Run Shiny App in a Background Process}
\usage{
runBackgroundApp(
  ui = NULL,
  server = NULL,
  appFile = NULL,
  appDir = NULL,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1"),
  stdout = "|",
  stderr = "|",
  env = NULL,
  options = list(),
  ...
)
}
\arguments{
\item{ui}{The UI definition of the app.}

\item{server}{A function with three parameters: \code{input}, \code{output}, and
\code{session}. The function is called once for each session ensuring that each
app is independent.}

\item{appFile}{Path to a .R file containing a Shiny application}

\item{appDir}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
\item An \code{.R} file containing a Shiny application, ending with an
expression that produces a Shiny app object.
\item A list with \code{ui} and \code{server} components.
\item A Shiny app object created by \code{\link[shiny:shinyApp]{shinyApp()}}.
}}

\item{port}{The TCP port that the application should listen on (defaults to 3000).}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{stdout}{The name of the file the standard output of
the child R process will be written to.
If the child process runs with the \code{--slave} option (the default),
then the commands are not echoed and will not be shown
in the standard output. Also note that you need to call \code{print()}
explicitly to show the output of the command(s).
IF \code{NULL} (the default), then standard output is not returned, but
it is recorded and included in the error object if an error happens.}

\item{stderr}{The name of the file the standard error of
the child R process will be written to.
In particular \code{message()} sends output to the standard
error. If nothing was sent to the standard error, then this file
will be empty. This argument can be the same file as \code{stdout},
in which case they will be correctly interleaved. If this is the
string \code{"2>&1"}, then standard error is redirected to standard output.
IF \code{NULL} (the default), then standard output is not returned, but
it is recorded and included in the error object if an error happens.}

\item{env}{Named character vector of environment variables to be passed}

\item{options}{Any options to be passed to shinyApp}

\item{...}{Extra arguments are passed to the \link[processx:process]{processx::process}
constructor.}
}
\value{
A \link[callr:r_process]{callr::r_process} object for the background Shiny app
}
\description{
Run Shiny App in a Background Process
}
\examples{
\dontrun{
app <- system.file("apps/sever-info-app.R", package = "shinybg")
bg_app <- runBackgroundApp(appFile = app, port = 3005, env = Sys.getenv()["USER"])
bg_app$kill()
}
}
