% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{view_app}
\alias{view_app}
\title{View background Shiny application}
\usage{
view_app(port)
}
\arguments{
\item{port}{TCP port the app is listening on (e.g., 3000)}
}
\value{
(Invisibly) The application's URL.
}
\description{
Open an application in an external browser or within the configured RStudio
viewer
}
\seealso{
Other app management functions: 
\code{\link{get_app}()},
\code{\link{kill_all_apps}()},
\code{\link{kill_app}()},
\code{\link{list_apps}()}
}
\concept{app management functions}
