% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBackgroundApp.R
\name{runBackgroundApp}
\alias{runBackgroundApp}
\title{Run Shiny App in a Background Process}
\usage{
runBackgroundApp(
  ui = NULL,
  server = NULL,
  appFile = NULL,
  appDir = NULL,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\arguments{
\item{ui}{The UI definition of the app.}

\item{server}{A function with three parameters: \code{input}, \code{output}, and
\code{session}. The function is called once for each session ensuring that each
app is independent.}

\item{appFile}{Path to a .R file containing a Shiny application}

\item{appDir}{The application to run. Should be one of the following:
\itemize{
\item A directory containing \code{server.R}, plus, either \code{ui.R} or
a \code{www} directory that contains the file \code{index.html}.
\item A directory containing \code{app.R}.
\item An \code{.R} file containing a Shiny application, ending with an
expression that produces a Shiny app object.
\item A list with \code{ui} and \code{server} components.
\item A Shiny app object created by \code{\link[shiny:shinyApp]{shinyApp()}}.
}}

\item{port}{The TCP port that the application should listen on (defaults to 3000).}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
}
\value{
A \link[callr:r_process]{callr::r_process} object for the background Shiny app
}
\description{
Run Shiny App in a Background Process
}
\examples{
\dontrun{
app <- system.file("apps/sever-info-app.R", package = "tiledbJupyterShiny")
bg_app <- runBackgroundApp(appFile = app, port = 3005)
bg_app$kill()
}
}
